♻️ refactor: OrderSagaState 클래스 경로 변경 및 코드 개선
- OrderSagaState 클래스를 기존 infrastructure/saga에서 infrastructure/saga/state 경로로 이동.
- 코드 내 관련 클래스 및 경로 수정.
- Order 및 OrderRequest를 기반으로 초기 상태를 생성하는 createInitialState 메서드 추가.
